+++ R/boilerplate-svm-poly.R
@@ -271,5 +271,5 @@ ts_auto_svm_poly <- function(.data, .date_col, .value_col, .formula, .rsamp_obj,
     attr(output, ".parsnip_engine") <- "kernlab"
     attr(output, ".function_family") <- "boilerplate"
 
-    return(invisible(output))
+    return(output)
 }